Refactor entity management in Doctrine component
The ServiceEntityRepository file was removed and replaced with SaveTrait and RemoveTrait files. These new traits encapsulate the shared logic previously held in ServiceEntityRepository providing clearer separation of concerns. Furthermore, an interface (IsEntityInterface) was created for stricter type control across the entities. These changes aim at improving clarity and robustness of the codebase. Updates were also made to composer.json to reflect the change in dependencies and the README file for installation and usage instructions. An MIT License file was added providing clarity on the permissible use of the software.